Edicao com auteticação de tipo de usuário

+++ Api/Rotas/Gerenciamento/Editar.ts
@@ -1,6 +1,7 @@
 import {Rotas} from '../Routes';
 import {Mongo} from '../../MongoDB/Mongo';
 import {Collections} from '../../MongoDB/MongoCollections';
+import {UsuarioService} from "../Usuario/usuarios.service";
 const multer = require('multer');
 
 import express = require('express');
@@ -29,24 +30,26 @@ const upload = multer({ storage: storage })
 
 
 app.put(Rotas.Cardapios, upload.single('file'), (req:any,res) =>{
-    console.log(req.body.Cardapio.Src);
+    console.log("token",req.body.token);
     try{
-        let query = 
-        {
-            Dia: req.body.Cardapio.Dia,
-            Nome: req.body.Cardapio.Nome,
-            Ingredientes: req.body.Cardapio.Ingredientes,
-            Tipo: req.body.Cardapio.Tipo,
-            Src: req.body.Cardapio.Src,
-            SrcType: req.body.Cardapio.SrcType
-        }
-        console.log("gerenciamento/"+Rotas.Cardapios,query)
+        UsuarioService.getByToken(req.body.token).then(user=>{
+            console.log(user);
+            if(user[0].Tipo == 2){
+                let query = 
+                {
+                    Dia: req.body.item.Cardapio.Dia,
+                    Nome: req.body.item.Cardapio.Nome,
+                    Ingredientes: req.body.item.Cardapio.Ingredientes,
+                    Tipo: req.body.item.Cardapio.Tipo,
+                    Src: req.body.item.Cardapio.Src,
+                    SrcType: req.body.item.Cardapio.SrcType
+                }
 
-        Mongo.Edit(Collections.Cardapio.NomeID, req.body.Cardapio._id, query).then(x=>{
-            redisConfig.flushAll();
-            console.log(x);
-
-            res.send(x);
+                Mongo.Edit(Collections.Cardapio.NomeID, req.body.item.Cardapio._id, query).then(x=>{
+                    redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
     }
     catch(err){
@@ -55,19 +58,23 @@ app.put(Rotas.Cardapios, upload.single('file'), (req:any,res) =>{
 });
 app.put(Rotas.InfoContato, (req:any,res) => {
     try{
+        UsuarioService.getByToken(req.body.token).then(user => {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                let query =
+                {
+                    Telefone: req.body.item.InformacoesContato.Telefone,
+                    Email: req.body.item.InformacoesContato.Email,
+                    HorarioAtendimento: req.body.item.InformacoesContato.HorarioAtendimento,
+                    Whatsapp: req.body.item.InformacoesContato.Whatsapp,
+                    Instagram: req.body.item.InformacoesContato.Instagram
+                }
 
-        let query = 
-        {
-            Telefone:req.body.InformacoesContato.Telefone,
-            Email:req.body.InformacoesContato.Email,
-            HorarioAtendimento:req.body.InformacoesContato.HorarioAtendimento,
-            Whatsapp:req.body.InformacoesContato.Whatsapp,
-            Instagram:req.body.InformacoesContato.Instagram
-        }
-
-        Mongo.Edit(Collections.InformacoesContato.NomeID, req.body.InformacoesContato._id, query).then(x=>{
-            redisConfig.flushAll();
-            res.send(x);
+                Mongo.Edit(Collections.InformacoesContato.NomeID, req.body.item.InformacoesContato._id, query).then(x => {
+                    redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
         
     }
@@ -77,21 +84,24 @@ app.put(Rotas.InfoContato, (req:any,res) => {
 });
 app.put(Rotas.Sobre, (req:any,res) => {
     try{
+        UsuarioService.getByToken(req.body.token).then(user => {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                let query =
+                {
+                    Descricao: req.body.item.Sobre.Descricao,
+                    Nome: req.body.item.Sobre.Nome,
+                    Servico: req.body.item.Sobre.Servico,
+                    Historia: req.body.item.Sobre.Historia,
+                    Slogan: req.body.item.Sobre.Slogan,
+                }
 
-        let query = 
-        {
-            Descricao:req.body.Sobre.Descricao,
-            Nome:req.body.Sobre.Nome,
-            Servico:req.body.Sobre.Servico,
-            Historia:req.body.Sobre.Historia,
-            Slogan:req.body.Sobre.Slogan,
-        }
-
-        Mongo.Edit(Collections.Sobre.NomeID, req.body.Sobre._id , query).then(x=>{
-            redisConfig.flushAll();
-            res.send(x);
+                Mongo.Edit(Collections.Sobre.NomeID, req.body.item.Sobre._id, query).then(x => {
+                    redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
-        
     }
     catch(err){
         res.send({erro:err});
@@ -99,14 +109,20 @@ app.put(Rotas.Sobre, (req:any,res) => {
 });
 app.put(Rotas.PrecoMarmitex, (req:any, res) => {
     try{
+        UsuarioService.getByToken(req.body.token).then(user => {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                let query =
+                {
+                    "Pequena": req.body.item.PrecoMarmitex.Pequena,
+                    "Grande": req.body.item.PrecoMarmitex.Grande
+                }
 
-        let query = 
-        {"Pequena": req.body.PrecoMarmitex.Pequena,
-         "Grande": req.body.PrecoMarmitex.Grande}
-        
-        Mongo.Edit(Collections.PrecoMarmitex.NomeID,req.body.PrecoMarmitex._id, query).then(x=>{
-            redisConfig.flushAll();
-            res.send(x);
+                Mongo.Edit(Collections.PrecoMarmitex.NomeID, req.body.item.PrecoMarmitex._id, query).then(x => {
+                    redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
         
     }
@@ -116,13 +132,18 @@ app.put(Rotas.PrecoMarmitex, (req:any, res) => {
 });
 app.put(Rotas.Complemento, (req: any, res) => {
     try{
-        let query = {"Nome": req.body.Complemento.Nome, "Tipo": req.body.Complemento.Tipo, "Preco": req.body.Complemento.Preco};
-        console.log(query);
-        Mongo.Filtrar(Collections.Complemento.NomeID, {"_id": req.body.Complemento._id}).then(result=>{
-            Mongo.EditarPorAtributo(Collections.Complemento.NomeID, {"Nome": req.body.Complemento.Nome}, query);
-            res.send({sucesso:req.body});
-            redisConfig.flushAll();
-        })
+        UsuarioService.getByToken(req.body.token).then(user => {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                let query = { "Nome": req.body.item.Complemento.Nome, "Tipo": req.body.item.Complemento.Tipo, "Preco": req.body.item.Complemento.Preco };
+                console.log(query);
+                Mongo.Filtrar(Collections.Complemento.NomeID, { "_id": req.body.item.Complemento._id }).then(result => {
+                    Mongo.EditarPorAtributo(Collections.Complemento.NomeID, { "Nome": req.body.item.Complemento.Nome }, query);
+                    res.send({ sucesso: req.body });
+                    redisConfig.flushAll();
+                })
+            }
+        });
     }
     catch(err){
         res.send({erro:err, query:req.body});
